+++ .github/workflows/tests.yml
@@ -39,7 +39,6 @@ jobs:
         test-subset:
           - |
             pymc/tests/test_util.py
-            pymc/tests/distributions/test_logprob.py
             pymc/tests/test_pytensorf.py
             pymc/tests/test_math.py
             pymc/tests/backends/test_base.py
