+++ pymc/tests/distributions/test_continuous.py
@@ -26,9 +26,10 @@ from pytensor.compile.mode import Mode
 
 import pymc as pm
 
-from pymc.distributions import logcdf, logp
 from pymc.distributions.continuous import Normal, get_tau_sigma, interpolated
 from pymc.distributions.dist_math import clipped_beta_rvs
+from pymc.logprob.abstract import logcdf
+from pymc.logprob.joint_logprob import logp
 from pymc.logprob.utils import ParameterValueError
 from pymc.pytensorf import floatX
 from pymc.tests.distributions.util import (
