+++ pymc/tests/distributions/test_multivariate.py
@@ -30,7 +30,6 @@ from pytensor.tensor.random.utils import broadcast_params
 
 import pymc as pm
 
-from pymc.distributions import logp
 from pymc.distributions.multivariate import (
     _LKJCholeskyCov,
     _OrderedMultinomial,
@@ -38,6 +37,7 @@ from pymc.distributions.multivariate import (
     quaddist_matrix,
 )
 from pymc.distributions.shape_utils import change_dist_size, to_tuple
+from pymc.logprob.joint_logprob import logp
 from pymc.logprob.utils import ParameterValueError
 from pymc.math import kronecker
 from pymc.pytensorf import compile_pymc, floatX, intX
