+++ pymc/tests/distributions/test_truncated.py
@@ -19,13 +19,14 @@ import scipy
 
 from pytensor.tensor.random.basic import GeometricRV, NormalRV
 
-from pymc import Censored, Model, draw, find_MAP, logp
+from pymc import Censored, Model, draw, find_MAP
 from pymc.distributions.continuous import Exponential, TruncatedNormalRV
 from pymc.distributions.shape_utils import change_dist_size
 from pymc.distributions.transforms import _default_transform
 from pymc.distributions.truncated import Truncated, TruncatedRV, _truncated
 from pymc.exceptions import TruncationError
 from pymc.logprob.abstract import _icdf
+from pymc.logprob.joint_logprob import logp
 from pymc.logprob.transforms import IntervalTransform
 from pymc.logprob.utils import ParameterValueError
 from pymc.tests.distributions.util import assert_moment_is_expected
